Refactored the vtk, split into multiple classes, not done yet
